requirements.txt
langchain
langchain-community
langgraph
trafilatura
playwright
google-api-python-client
google-auth
google-auth-oauthlib
gspread
openai
anthropic
python-dotenv 
mcp
mcp-client
mcp-server
mcp-tools
mcp-tools-bsky
mcp-tools-google-sheets
mcp-tools-twitter
aiohttp